Refactor hopping code into DefaultScheduler class in scheduler.py
